deplace utils functions and const in other file
